Changed double to unsigned for type of nSubsets variable in UnimapEdgeMove::update to eliminate warning
git-svn-id: https://phycas.svn.sourceforge.net/svnroot/phycas/trunk@1273 b774380d-c90d-0410-997c-aec342e5d6fd

+++ phycas/src/unimap_edge_move.cpp
@@ -224,7 +224,7 @@ bool UnimapEdgeMove::update()
 
 	proposeNewState();
 	PartitionModelShPtr partModel = likelihood->getPartitionModel();
-	const double nSubsets = partModel->getNumSubsets();
+	const unsigned nSubsets = partModel->getNumSubsets();
 	std::vector<double> uniformization_lambda(nSubsets);
 	for (unsigned i = 0; i < nSubsets; ++i)
 		{
